backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
redis==5.2.1
python-multipart==0.0.20
pydantic-settings==2.8.1
